+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* `Super::Reset`. Include it in your controller to remove all the controller
+  methods and actions that Super defines
+
 ### Removed
 
 * The README (cheatsheet) that was added when installing Super for the first